requirements.txt
gradio>=5.0
google-generativeai
python-dotenv
